+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+- fix android monitoring issues when minor and/or minor where null
+- add monitoring in example
+
 ## 1.0.1
 README and config fixes